+++ packages/orq-rc/RELEASES.md
@@ -5899,3 +5899,13 @@ Based on:
 - [typescript v4.2.0-rc.36] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.36 - packages/orq-rc
+
+## 2026-01-17 09:53:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.687.1 (2.796.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.37] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.37 - packages/orq-rc